feat: code adjustment+error handling no data key

+++ backend/deamon/tcp_server.py
@@ -53,18 +53,24 @@ class Tcpsocket:
             conn, addr = self.server.accept()
             tmp = conn.recv(1024)
             data = json.loads(tmp)
-            cursor = Database.find_ip("Ip_whitelist")
+            keysList = list(data.keys())
+            cursor = Database.find_ip("Ip_allowlist")
             ip_adresses = cursor["Ip_Adresses"]
-            if addr[0] in ip_adresses:
-                if Database.countDocument("Errorlog", {"errorcode": 42, "machine": data['data'][0]}) > 0:
-                    Database.deleteOne("Errorlog", {"errorcode": 42, "machine": data['data'][0]})
-                print(f"Connected by {addr}")
-                thread = threading.Thread(target=self.handle_client, args=(conn, data_handler, addr))
-                thread.start()
+            if "data" in keysList:
+                if addr[0] in ip_adresses:
+                    if Database.countDocument("Errorlog", {"errorcode": 42, "machine": data['data'][0]}) > 0:
+                        Database.deleteOne("Errorlog", {"errorcode": 42, "machine": data['data'][0]})
+                    print(f"Connected by {addr}")
+                    thread = threading.Thread(target=self.handle_client, args=(conn, data_handler, addr))
+                    thread.start()
+                else:
+                    print("Wrong Ip")
+                    Database.replace("Errorlog", {"errorcode": 42, "errormsg": f"Cannot connect, {addr[0]} not in allowlist",
+                    "machine": data['data'][0]}, {"machine": data['data'][0], "errorcode": 42})
+                    conn.shutdown(socket.SHUT_RDWR)
+                    conn.close()
             else:
-                print("Wrong Ip")
-                Database.replace("Errorlog", {"errorcode": 42, "errormsg": f"Cannot connect, {addr[0]} not in allowlist",
-                "machine": data['data'][0]}, {"machine": data['data'][0], "errorcode": 42})
+                logging.debug(f"No data key in JSON Object Machine with ip: {addr[0]} rejected")
                 conn.shutdown(socket.SHUT_RDWR)
                 conn.close()
